Struct aws_sdk_codeartifact::input::CopyPackageVersionsInput

#[non_exhaustive]
pub struct CopyPackageVersionsInput {
    pub domain: Option<String>,
    pub domain_owner: Option<String>,
    pub source_repository: Option<String>,
    pub destination_repository: Option<String>,
    pub format: Option<PackageFormat>,
    pub namespace: Option<String>,
    pub package: Option<String>,
    pub versions: Option<Vec<String>>,
    pub version_revisions: Option<HashMap<String, String>>,
    pub allow_overwrite: Option<bool>,
    pub include_from_upstream: Option<bool>,
}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

domain: Option<String>

The name of the domain that contains the source and destination repositories.

domain_owner: Option<String>

The 12-digit account number of the AWS account that owns the domain. It does not include dashes or spaces.

source_repository: Option<String>

The name of the repository that contains the package versions to copy.

destination_repository: Option<String>

The name of the repository into which package versions are copied.

format: Option<PackageFormat>

The format of the package that is copied. The valid package types are:

npm: A Node Package Manager (npm) package.

pypi: A Python Package Index (PyPI) package.

maven: A Maven package that contains compiled code in a distributable format, such as a JAR file.

namespace: Option<String>

The namespace of the package. The package component that specifies its namespace depends on its type. For example:

The namespace of a Maven package is its groupId.

The namespace of an npm package is its scope.

A Python package does not contain a corresponding component, so Python packages do not have a namespace.

package: Option<String>

The name of the package that is copied.

versions: Option<Vec<String>>

The versions of the package to copy.

You must specify versions or versionRevisions. You cannot specify both.

version_revisions: Option<HashMap<String, String>>

A list of key-value pairs. The keys are package versions and the values are package version revisions. A CopyPackageVersion operation succeeds if the specified versions in the source repository match the specified package version revision.

You must specify versions or versionRevisions. You cannot specify both.

allow_overwrite: Option<bool>

Set to true to overwrite a package version that already exists in the destination repository. If set to false and the package version already exists in the destination repository, the package version is returned in the failedVersions field of the response with an ALREADY_EXISTS error code.

include_from_upstream: Option<bool>

Set to true to copy packages from repositories that are upstream from the source repository to the destination repository. The default setting is false. For more information, see Working with upstream repositories.

Implementations

impl CopyPackageVersionsInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<CopyPackageVersions, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<CopyPackageVersions>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture CopyPackageVersionsInput