Module aws_sdk_codeartifact::model

Modules

asset_summary

See AssetSummary

domain_description

See DomainDescription

domain_summary

See DomainSummary

license_info

See LicenseInfo

package_dependency

See PackageDependency

package_summary

See PackageSummary

package_version_description

See PackageVersionDescription

package_version_error

See PackageVersionError

package_version_summary

See PackageVersionSummary

repository_description

See RepositoryDescription

repository_external_connection_info

See RepositoryExternalConnectionInfo

repository_summary

See RepositorySummary

resource_policy

See ResourcePolicy

successful_package_version_info

See SuccessfulPackageVersionInfo

tag

See Tag

upstream_repository

See UpstreamRepository

upstream_repository_info

See UpstreamRepositoryInfo

Structs

AssetSummary

DomainDescription

DomainSummary

Information about a domain, including its name, Amazon Resource Name (ARN), and status.

LicenseInfo

PackageDependency

PackageSummary

PackageVersionDescription

PackageVersionError

PackageVersionSummary

RepositoryDescription

The details of a repository stored in AWS CodeArtifact. A CodeArtifact repository contains a set of

RepositoryExternalConnectionInfo

RepositorySummary

Details about a repository, including its Amazon Resource Name (ARN), description, and

ResourcePolicy

SuccessfulPackageVersionInfo

Tag

A tag is a key-value pair that can be used to manage, search for, or filter resources in AWS CodeArtifact.

UpstreamRepository

UpstreamRepositoryInfo

Enums

DomainStatus

ExternalConnectionStatus

HashAlgorithm

PackageFormat

PackageVersionErrorCode

PackageVersionSortType

PackageVersionStatus

ResourceType

ValidationExceptionReason