Module aws_sdk_codeartifact::operation

Structs

AssociateExternalConnection

Adds an existing external connection to a repository. One external connection is allowed

CopyPackageVersions

CreateDomain

CreateRepository

DeleteDomain

DeleteDomainPermissionsPolicy

DeletePackageVersions

Deletes one or more versions of a package. A deleted package version cannot be restored

DeleteRepository

DeleteRepositoryPermissionsPolicy

DescribeDomain

DescribePackageVersion

DescribeRepository

DisassociateExternalConnection

DisposePackageVersions

GetAuthorizationToken

GetDomainPermissionsPolicy

GetPackageVersionAsset

GetPackageVersionReadme

GetRepositoryEndpoint

GetRepositoryPermissionsPolicy

ListDomains

Returns a list of DomainSummary objects for all domains owned by the AWS account that makes

ListPackageVersionAssets

ListPackageVersionDependencies

ListPackageVersions

ListPackages

ListRepositories

ListRepositoriesInDomain

ListTagsForResource

Gets information about AWS tags for a specified Amazon Resource Name (ARN) in AWS CodeArtifact.

PutDomainPermissionsPolicy

PutRepositoryPermissionsPolicy

TagResource

Adds or updates tags for a resource in AWS CodeArtifact.

UntagResource

Removes tags from a resource in AWS CodeArtifact.

UpdatePackageVersionsStatus

UpdateRepository